feat(models): add InstanceNorm up-down iso model for 8nm training
updown_act_iso.py's group == 0 branch builds rsunet_act (BatchNorm), while
rsunet_act_iso.py builds rsunet_act_in (InstanceNorm). Since no config sets
--group and it defaults to 0, switching an rsunet_act_iso experiment to the
up-down wrapper silently changes the normalization: loading a 260805-style
checkpoint leaves 87 BatchNorm buffers (29 norm layers x running_mean /
running_var / num_batches_tracked) uninitialized under --pretrain, and fails
outright on a strict resume.

updown_act_iso_in.py is updown_act_iso.py with that branch switched to
rsunet_act_in. DownBlock and UpBlock hold no parameters and 'in' / 'core' /
'out' keep their names, so the state_dict matches rsunet_act_iso exactly.
Verified against 260805-SR-finetune-invertebrate-aff-head-x0/model300000.chkpt
at fov 160 / outputsz 120: 102/102 model tensors load, 0 missing, 0 shape
mismatches, and the only ignored checkpoint keys are the dropped myelin head.

Also:

- aug_8nm.py: the isotropic branch of super_resolution/aug_v0.py with the
  occlusion boxes doubled (dims 5-25 -> 10-50, margin 5 -> 10) so they keep
  their physical size when the voxel size halves from 16nm to 8nm. Note blur
  is MixedBlurrySection(maxsec=...), an absolute section count bounded by the
  patch (Section._validate asserts zmax > maxsec), so it stays a config knob
  rather than being scaled here.

- multi_zettaset: read the image as float32. `convert_array(vol) / 255.0`
  promoted uint8 to float64, and postprocess casts to float32 anyway. At 8nm
  a padded hemibrain sample is ~314M voxels, so this is 1.2 GB per sample --
  7 GB across the six training samples.

+++ deepem/data/dataset/multi_zettaset.py
@@ -227,7 +227,9 @@ def load_sample(
         fill_missing=True,
         bounded=False,
     )[image_bbox.to_slices()]
-    dset["input"] = convert_array(vol) / 255.0
+    # float32, not the float64 that uint8 / 255.0 would promote to: at 8 nm a
+    # padded hemibrain sample is ~314 M voxels, so the dtype is worth 1.3 GB.
+    dset["input"] = convert_array(vol).astype(np.float32) / np.float32(255.0)
     print(f"\tinput: {dset['input'].shape}")
 
     # Assumes that zettaset's annotation names follow DeepEM's convention.
